+++ temp/sb3o/optimizations/index.js
@@ -1,0 +1,9 @@
+module.exports = {
+  'remove-meta': require('./remove-meta'),
+  'optimize-ids': require('./optimize-ids'),
+  'remove-redundant-monitor-value': require('./remove-redundant-monitor-value'),
+  'optimize-blocks': require('./optimize-blocks'),
+  'optimize-mutations': require('./optimize-mutations'),
+  'optimize-svg': require('./optimize-svg'),
+  'optimize-png': require('./optimize-png')
+};