+++ temp/sb3o/optimizations/optimization.js
@@ -1,0 +1,15 @@
+class Optimization {
+  constructor(deserializedProject) {
+    this.project = deserializedProject;
+  }
+
+  warn(...args) {
+    console.warn(`${this.constructor.name}: warn:`, ...args);
+  }
+
+  run() {
+    throw new Error('Not implemented');
+  }
+}
+
+module.exports = Optimization;